[FEAT/#69] 체크리스트 API, 프로필 이미지 불러오기

+++ app/src/main/java/com/umc/timeCAlling/data/dto/response/schedule/SuccessRateResponseDto.kt
@@ -6,7 +6,7 @@ data class SuccessRateResponseDto(
     val total: Int,
     val success: Int,
     val failed: Int,
-    val successRate: Int,
+    val successRate: Float,
 ) {
     fun toSuccessRateResponseModel() = SuccessRateResponseModel(total, success, failed, successRate)
 }
